+++ src/NEventStore.Persistence.Sql/SqlPersistenceEngine.Async.cs
@@ -259,9 +259,9 @@ namespace NEventStore.Persistence.Sql
 				string statement = _dialect.GetStreamsRequiringSnapshots;
 				query.AddParameter(_dialect.BucketId, bucketId, DbType.AnsiString);
 				query.AddParameter(_dialect.Threshold, maxThreshold);
-				query.AddParameter(_dialect.StreamId, String.Empty, DbType.AnsiString); // (NEventStore.Persistence.Sql #54) to define the parameter for paging
+				query.AddParameter(_dialect.StreamId, _dialect.CoalesceParameterValue(String.Empty), DbType.AnsiString); // (NEventStore.Persistence.Sql #54) to define the parameter for paging
 				return query.ExecutePagedQueryAsync(
-					statement, (q, s) => q.SetParameter(_dialect.StreamId, _dialect.CoalesceParameterValue(s.StreamId()), DbType.AnsiString), // todo: I'm not sure this is used, the query does not have a "StreamId" parameter
+					statement, (q, s) => q.SetParameter(_dialect.StreamId, _dialect.CoalesceParameterValue(s.StreamIdHash()), DbType.AnsiString),
 					asyncObserver, cancellationToken);
 			},
 			adapter, cancellationToken);